+++ math.h
@@ -1,4 +1,4 @@
 int somma(int, int);
 int differenza(int, int);
 int moltiplicazione(int, int);
-int divisione(int, int);
+float divisione(float, float);
